moved home and profile to pages folder for better organization
